+++ .travis.yml
@@ -3,8 +3,6 @@ language: go
 matrix:
   fast_finish: true
   include:
-  - go: 1.6.x
-  - go: 1.7.x
   - go: 1.8.x
   - go: 1.9.x
   - go: 1.10.x
