+++ version.go
@@ -5,4 +5,4 @@
 package gin
 
 // Version is the current gin framework's version.
-const Version = "v1.4.0-dev"
+const Version = "v1.4.0"
